pull out BaseStudentView
