+++ carr/mixins.py
@@ -1,7 +1,9 @@
 import json
 
 from django.contrib.auth.decorators import login_required
+from django.contrib.auth.models import User
 from django.http import HttpResponse
+from django.shortcuts import get_object_or_404, render
 from django.utils.decorators import method_decorator
 from django.views.generic.base import View
 
@@ -48,3 +50,25 @@ class BaseSaveStateView(View):
         response['success'] = 1
 
         return HttpResponse(json.dumps(response), 'application/json')
+
+
+class BaseStudentView(View):
+    template_name = None  # these must be overridden
+    state_class = None
+
+    def get_student_user(self, request, user_id):
+        if request.user.user_type() == "student":
+            return request.user
+        else:
+            return get_object_or_404(User, id=user_id)
+
+    def get_context_data(self, request, user_id):
+        student_user = self.get_student_user(request, user_id)
+        return {
+            'student': student_user,
+            'student_json': state_json(self.state_class, student_user)
+        }
+
+    def get(self, request, user_id):
+        return render(request, self.template_name,
+                      self.get_context_data(request, user_id))
